app/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dateutil==2.9.0
email-validator>=2.0.0
python-multipart>=0.0.9

asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.0.0
httpx>=0.27.0
